update chatta demo up to stable geotrellis; spray -> akka http update
