+++ geotrellis/ingest.sh
@@ -1,19 +1,15 @@
 #!/usr/bin/env bash
 
 # GeoTrellis ingest jar
-export JAR="target/scala-2.10/GeoTrellis-Tutorial-Project-assembly-0.1-SNAPSHOT.jar"
+export JAR="target/scala-2.11/geotrellis-chatta-demo-assembly-0.1-SNAPSHOT.jar"
 
-# Remove some bad signatures from the assembled JAR
-zip -d $JAR META-INF/ECLIPSEF.RSA > /dev/null
-zip -d $JAR META-INF/ECLIPSEF.SF > /dev/null
-
-echo "--class geotrellis.chatta.ChattaIngest --driver-memory=2G target/scala-2.10/GeoTrellis-Tutorial-Project-assembly-0.1-SNAPSHOT.jar \
+echo "--class geotrellis.chatta.ChattaIngest --driver-memory=2G $JAR \
   --input file:///${PWD}/conf/input.json \
   --output file://${PWD}/conf/output.json \
   --backend-profiles file://${PWD}/conf/backend-profiles.json"
 
 spark-submit \
-  --class geotrellis.chatta.ChattaIngest --driver-memory=2G target/scala-2.10/GeoTrellis-Tutorial-Project-assembly-0.1-SNAPSHOT.jar \
+  --class geotrellis.chatta.ChattaIngest --driver-memory=2G $JAR \
   --input "file:///${PWD}/conf/input.json" \
   --output "file://${PWD}/conf/output.json" \
   --backend-profiles "file://${PWD}/conf/backend-profiles.json"
