+++ geotrellis/src/main/scala/geotrellis/chatta/Main.scala
@@ -1,12 +1,28 @@
 package geotrellis.chatta
 
+import geotrellis.kryo.AvroRegistrator
+
 import akka.actor.Props
 import akka.io.IO
 import akka.actor.ActorSystem
 import com.typesafe.config.ConfigFactory
-import spray.can.Http
+import akka.actor.{ActorRef, ActorSystem, Props}
+import akka.event.Logging
+import akka.http.scaladsl.Http
+import akka.stream.ActorMaterializer
+import org.apache.spark.{SparkConf, SparkContext}
 
-object Main {
+object AkkaSystem {
+  implicit val system = ActorSystem("chatta-demo")
+  implicit val materializer = ActorMaterializer()
+
+  trait LoggerExecutor {
+    protected implicit val log = Logging(system, "app")
+  }
+}
+
+object Main extends ChattaServiceRouter {
+  import AkkaSystem._
 
   val weights = Map(
     "ImperviousSurfacesBarrenLandsOpenWater" -> 1,
@@ -21,21 +37,23 @@ object Main {
     "FarmlandOrForestedLandsWithPrimeAgriculturalSoils" -> 10
   )
 
+  val config = ConfigFactory.load()
+  val staticPath = config.getString("geotrellis.server.static-path")
+  val port = config.getInt("geotrellis.port")
+  val host = config.getString("geotrellis.hostname")
+
+  val conf = AvroRegistrator(
+    new SparkConf()
+      .setAppName("ChattaDemo")
+      .set("spark.serializer", classOf[org.apache.spark.serializer.KryoSerializer].getName)
+      .set("spark.kryo.registrator", classOf[geotrellis.spark.io.kryo.KryoRegistrator].getName)
+  )
+
+  implicit val sc = new SparkContext(conf)
+
+  lazy val (reader, tileReader, attributeStore) = initBackend(config)
+
   def main(args: Array[String]): Unit = {
-
-    implicit val system = ActorSystem("chatta-demo")
-
-    val config = ConfigFactory.load()
-    val staticPath = config.getString("geotrellis.server.static-path")
-    val port = config.getInt("geotrellis.port")
-    val host = config.getString("geotrellis.hostname")
-
-    // create and start our service actor
-    val service = system.actorOf(Props(classOf[ChattaServiceActor], staticPath, config), "chatta-service")
-
-    // start a new HTTP server on port 8080 with our service actor as the handler
-    IO(Http) ! Http.Bind(service, host, port)
+    Http().bindAndHandle(routes, host, port)
   }
-
 }
-
